Fix race condition between the "debounce" and "update" timeouts
When a script with an update interval of one second was removed
from the directory, it would take one second for the debounce
to destroy its button. Meanwhile, another update could happen,
which would fail as the file did not exist anymore.
The solution is to immediately destroy (invalidate) all buttons
on every "changed" event and only debounce *adding* buttons.