Close the standard output stream instead of the file descriptor
This fixes a mysterious bug where one in every few hundred calls
to read_line_finish would fail.
While difficult to verify, the theory is that the stream closes
the file descriptor when it is garbage collected, which could
happen *after* a new process reusing the same numerical file
descriptor had been spawned, causing the active pipe to be closed
instead of the old one.