menuitem.js: Backward-compatible GNOME 3.36 support
This makes the extension compatible with GNOME 3.36 without giving up
support for older GNOME versions. Tested with 3.26, 3.34, 3.36; should run
on older versions as well.

Rationale for the implementation: there are two basic types of classes
in GNOME shell: plain ES6 classes, and classes that extend Gobject and
are created with GObject.registerClass(). Sometimes a plain class is
converted to a GObject class in a later GJS version. In this case, the
syntax for derived classes has to be converted from plain to GObject, too.
In extensions that need to support a broader range of GNOME versions,
these two class definitions can't be easily combined with conditional
code. The main reason is the different initialization of the object
itself (constructor vs. _init() and the base object (super() vs.
super._init()).

However, it's possible to separate the actual implementation from the
class definition, and use bind() to make access "this" in these functions
in different ways, depending on the way the class is defined.

For simple classes with just one method (constructor), such as
ArgosMenuItem, this can be done using a generic function.